Remove one more hardcoded `master` reference
Removing the one remaining hardcoded references to `master` that I discovered.

Originally I assumed we'd need to flip this to `HEAD`, but in
`https://github.com/dependabot/dependabot-core/pull/6149` we were able to
get away with simply removing the hardcoded reference altogether in the
`bundler` ecosystem and we haven't seen any ill effects the past 6 months...

So let's try removing this and see if it works as smoothly in these
ecosystems as it did in the `bundler` ecosystem.